Move the app to investigations directory
