+++ docs/internal/investigations/runtime-director-dapr/component/internal/store/store.go
@@ -6,8 +6,8 @@ import (
 	"path/filepath"
 
 	"github.com/ghodss/yaml"
-	"github.com/kyma-project/control-plane/poc/component/pkg/apperror"
-	"github.com/kyma-project/control-plane/poc/component/pkg/model"
+	"github.com/kyma-project/control-plane/docs/internal/investigations/runtime-director-dapr/component/pkg/apperror"
+	"github.com/kyma-project/control-plane/docs/internal/investigations/runtime-director-dapr/component/pkg/model"
 )
 
 type storeConfig struct {
